Change "clear log content" API endpoint to "set log content"
Clearing a log can be implemented by setting the log to the empty
string. This API endpoint is more versatile, though, and can be used for
things like writing an error to the logs.

The new endpoint also takes the job ID instead of the log ID, since the
log ID is an internal implementation detail that shouldn't really leak
outside of the logs app.