Speed up test suite significantly
The retrying code within ArchiveLog was executing within our test
environment, which was slowing things down unnecessarily.

The backoff is clever in that it simply adds a second for each loop
(defaulting to 5 times), but it adds (1 + 2 + 3 + 4) seconds to the test
suite.

I'm not sure if this is our preferred approach (maybe stubbing a
constant would be nicer), but it's effective.