Don't update aggregated_at nor content if there're no log parts (#156)
When the aggregation process runs we first gather all of the ids to
aggregate and then we run aggregate code on each of the logs. If a user
restarts a job after the ids are fetched, but before the log for the job
was aggregated, the restart will clean log parts. In such case we would
set content to NULL and aggregated_at to current time.

This commit changes the query to aggregate log to rerun log_parts check,
meaning that the UPDATE will not update anything unless a log with parts
is available.