remove constraint on log_id
the theory is that this may be causing maintenance to
take a long time, as postgres needs to scan the table
to make sure the constraint applies to all rows in
the partition.

we are not gaining much from these constraints, so
it should be relatively safe to drop them -- we may
incur a bit more i/o for index scans on one more
partition.