support rabbitmq sharding plugin
In order to support rabbitmq-sharding, we need to make sure
to skip declaring the sharded queue. The sharding plugin only
works if the queue does not exist, it will then dynamically
assign the virtual queue to one of the shards for each consumer.

This also switches between `reporting.jobs.logs` and
`reporting.jobs.logs_sharded` as queue names depending on
whether sharding is enabled.